Initial commit of Python look-up performance test
Initial commit of Python look-up table performance test.  The idea is to
answer the following question:

    Is it faster to create and store a lookup table to translate a tuple
    to a one-dimensional flattened index, or is it faster to calculate
    the flattened index?

To test this, we create a lookup table (i.e., dictionary) mapping each
permutation of 34x34 to a one-dimensional index value, calculated as:

    tuple[0] * 34 + tuple[1]

To push the experiment further, a Cython version is also included for
comparison.